+++ app/src/main/res/layout/activity_archive.xml
@@ -8,7 +8,7 @@
                 android:paddingLeft="16dp"
                 android:paddingRight="16dp"
                 android:paddingTop="16dp"
-                tools:context="pizzarat.cs2340.gatech.edu.ratpatrol.ArchiveActivity">
+    tools:context="pizzarat.cs2340.gatech.edu.ratpatrol.archive.ArchiveActivity">
 
     <SearchView
         android:id="@+id/searchView"
