+++ app/src/main/res/values/strings.xml
@@ -39,4 +39,7 @@
     <string name="navigation_drawer_close">Close navigation drawer</string>
 
     <string name="action_settings">Settings</string>
+
+    <string name="search_label">Search Address...</string>
+    <string name="search_hint">Enter some text</string>
 </resources>
